vae_latent/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
tensorboard>=2.13.0
PyYAML>=6.0
Pillow>=9.0.0
tqdm>=4.65.0
einops>=0.6.0

